Add a class to read a team config and automatically reload it on change.
Reloading is useful for allowing last minute changes, adding teams or changing names. A reload does not push the changes out to clients, as it is expected that changes will only be used for teams not currently playing. It is possible to force updates to clients, e.g. by reloading the client or toggling current team names temporarily.