Add a team picture overlay view at /overlay/teams.
Pictures must be in the public/photos directory, and match the player name exactly with a .jpg extension. There is no handling of a missing photo.
Teams should be set up using the teams.conf file, so that players can be added. If player names are not known, the image boxes will be left empty. Players are shown in the same order as in the configuration file.